fix: bedrock-kb-retrieval-mcp-server hardcoded region (#86)
* fix: adds AWS_REGION environment variable in client initialization.

Updated the AWS client initialization in server.py to include region_name and os.getenv(). This change ensures that both the runtime and management clients are instantiated with the specified AWS region, improving the flexibility of the application.

* chore: remove Dockerfile and Smithery configuration for nova-canvas-mcp-server

This commit deletes the Dockerfile and the Smithery configuration file for the nova-canvas-mcp-server until we figure out listing procedures

* updates typing